+++ src/packages/piral-lazy/src/types.ts
@@ -8,6 +8,13 @@ declare module 'piral-core/lib/types/custom' {
   }
 }
 
+export interface DependencyCache {
+  [name: string]: {
+    result: Promise<any>;
+    loader: LazyDependencyLoader;
+  };
+}
+
 export interface LazyComponentLoader<TProps> {
   (): Promise<FirstParametersOf<ComponentConverters<TProps>>>;
   current?: Promise<ForeignComponent<TProps>>;
@@ -19,9 +26,17 @@ export interface LazyComponent<TProps> {
    */
   load: LazyComponentLoader<TProps>;
   /**
+   * The promises resolving in the optional dependencies.
+   */
+  deps?: Array<Promise<any>>;
+  /**
    * The type of the lazy component.
    */
   type: 'lazy';
+}
+
+export interface LazyDependencyLoader {
+  (): Promise<any>;
 }
 
 /**
@@ -29,9 +44,16 @@ export interface LazyComponent<TProps> {
  */
 export interface PiletLazyApi {
   /**
+   * Defines a dependency for lazy loading.
+   * @param name The name of the dependency.
+   * @param loader The associated dependency loader.
+   */
+  defineDependency(name: string, loader: LazyDependencyLoader): void;
+  /**
    * Properly introduces a lazy loaded foreign component.
    * @param cb The callback to trigger when the component should be loaded.
+   * @param deps The optional names of the dependencies to load beforehand.
    * @returns The lazy loading component.
    */
-  fromLazy<T>(cb: LazyComponentLoader<T>): LazyComponent<T>;
+  fromLazy<T>(cb: LazyComponentLoader<T>, deps?: Array<string>): LazyComponent<T>;
 }