+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Updated documentation page (#206)
 * Added `once` flag for scaffolding (#225)
 * Fixed potentially non-working Piral CLI scaffold on Windows (#192)
+* Added `defineDependency` to `piral-lazy` (#226)
 
 ## 0.11.5 (May 24, 2020)